Improve LC download robustness & numeric precision
Add widespread float64 guards and robustness fixes across the pipeline to preserve shallow transit signals and harden downloads and vetting. Key changes include: BLS search: new scan_depth parameter, adaptive grid density and tighter transit masking; Detrending: cast arrays to float64, asymmetric sigma-clipping of positive outliers, stellar-radius-based window scaling and use of cleaned flux; Vetting: accept an snr parameter and treat high-SNR signals as "Likely Planet"; LightkurveClient: precision policy docstring, S3 anonymous fallback, safer temp-file streaming, stricter size checks, increased HTTP timeouts, a dedicated multi-sector TESS download_and_validate path (download_all with per-sector validation and normalization), and final dtype casts on stitched outputs; Orchestrator: ensure input arrays are float64. Also add streamlit dark theme, QA scripts/tests, logs and UI test outputs to repository. These changes improve detection fidelity for small transits and make data acquisition resilient to network failures and corrupt files.

+++ astraeus/core/orchestrator.py
@@ -104,15 +104,15 @@ def run_multi_planet_search(raw_lightcurve, max_signals=5, snr_floor=7.1):
     """
     # Extract time and flux from raw_lightcurve
     if isinstance(raw_lightcurve, dict):
-        time = np.asarray(raw_lightcurve.get('time', []))
-        flux = np.asarray(raw_lightcurve.get('flux', []))
+        time = np.asarray(raw_lightcurve.get('time', []), dtype=np.float64)
+        flux = np.asarray(raw_lightcurve.get('flux', []), dtype=np.float64)
         target_name = raw_lightcurve.get('target_name', 'Unknown')
         data_source = raw_lightcurve.get('data_source', 'Unknown')
         metadata = raw_lightcurve.get('metadata', {})
     else:
         # Fallback assuming object notation
-        time = np.asarray(getattr(raw_lightcurve, 'time', []))
-        flux = np.asarray(getattr(raw_lightcurve, 'flux', []))
+        time = np.asarray(getattr(raw_lightcurve, 'time', []), dtype=np.float64)
+        flux = np.asarray(getattr(raw_lightcurve, 'flux', []), dtype=np.float64)
         target_name = getattr(raw_lightcurve, 'target_name', 'Unknown')
         data_source = getattr(raw_lightcurve, 'data_source', 'Unknown')
         metadata = getattr(raw_lightcurve, 'metadata', {})
